Used PSR-4 instead of PSR-0 Changed required php version to 5.3 Removed color related methods Removed set/ clear methods Changed namespace to Gemabit\String
