+++ .gitignore
@@ -1,2 +1,3 @@
 vendor/*
 *.lock
+tests/index.php
