+++ composer.json
@@ -2,20 +2,21 @@
     "name": "gemabit/string",
     "type": "library",
     "description": "A super string object that'll make your life easier.",
-    "keywords": ["string","helper"],
+    "keywords": ["string", "helper"],
     "license": "Apache-2.0",
     "authors": [
         {
             "name": "André Filipe",
-            "email": "andre.r.flip@gmail.com"
+            "email": "andre.r.flip@gmail.com",
+            "role": "Developer"
         }
     ],
     "require": {
-        "php": ">=5.4"
+        "php": ">=5.3"
     },
     "autoload": {
-        "psr-0": {
-            "Gemabit\\Tools\\String\\": "src/"
+        "psr-4": {
+            "Gemabit\\String\\": "src/"
         }
     }
 }
